initial lesp grammar
We specify the syntax for lesp, a simple s-expression/lisp dialect.
It is intended as a structured markup language, with more defined
nesting than unstructured formats like Markdown or Asciidoc, but less
verbosity than XML-based formats like HTML or Docbook.

There is only one real scalar value, strings, which
do *not* need to be quoted, because lesp is intended as a way to write
structured texts. Strings cannot contain whitespace, it is implied on
the meta-level (i.e. two strings would be separated by a simple space
by default, and linebreaks are introduced by wrapping in paragraph
elements).

S-Expressions form the structural backbone, their first element is
always a string, and has special meaning on the meta-level. Further
elements can be arbitrary expressions.

The only other syntactic elements are lists (angled brackets), which
are flat lists of strings (no nested expressions allowed) and dicts,
which are mappings of keys (strings) to values (also strings).

Example sexp: (p this (is a (nested)) sexp)
Example list: [ a b c ]
Example dict: { foo=bar bar = baz }

The only structural characters are therefor ([{=}]}), which can all
be escaped by \. \ can be escaped by \\.